7z: fix unpacking error when trying to (unsuccessfully) unpack encrypted 7z files

+++ src/bang/parsers/archivers/7z/UnpackParser.py
@@ -126,25 +126,26 @@ class SevenzipUnpackParser(UnpackParser):
         pass
 
     def unpack(self, meta_directory):
-        # walk the results directory
-        for result in self.unpack_directory.glob('**/*'):
-            # first change the permissions
-            result.chmod(stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR)
+        if not self.encrypted:
+            # walk the results directory
+            for result in self.unpack_directory.glob('**/*'):
+                # first change the permissions
+                result.chmod(stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR)
 
-            file_path = result.relative_to(self.unpack_directory)
+                file_path = result.relative_to(self.unpack_directory)
 
-            if result.is_symlink():
-                meta_directory.unpack_symlink(file_path, result.readlink())
-            elif result.is_dir():
-                meta_directory.unpack_directory(file_path)
-            elif result.is_file():
-                with meta_directory.unpack_regular_file_no_open(file_path) as (unpacked_md, outfile):
-                    self.local_copy2(result, outfile)
-                    yield unpacked_md
-            else:
-                continue
+                if result.is_symlink():
+                    meta_directory.unpack_symlink(file_path, result.readlink())
+                elif result.is_dir():
+                    meta_directory.unpack_directory(file_path)
+                elif result.is_file():
+                    with meta_directory.unpack_regular_file_no_open(file_path) as (unpacked_md, outfile):
+                        self.local_copy2(result, outfile)
+                        yield unpacked_md
+                else:
+                    continue
 
-        shutil.rmtree(self.unpack_directory)
+            shutil.rmtree(self.unpack_directory)
 
     # a wrapper around shutil.copy2 to copy symbolic links instead of
     # following them and copying the data.
